refactor: type들 따로 파일로 분리해서 관리
